Move console selection to ExternalTab.java
