+++ src/main/java/net/sf/jabref/gui/preftabs/ExternalTab.java
@@ -16,15 +16,28 @@
 package net.sf.jabref.gui.preftabs;
 
 import java.awt.BorderLayout;
+import java.awt.GridBagConstraints;
+import java.awt.GridBagLayout;
 import java.awt.GridLayout;
+import java.awt.Insets;
+import java.awt.event.ActionEvent;
+import java.awt.event.ActionListener;
+import java.nio.file.Files;
+import java.nio.file.Path;
+import java.nio.file.Paths;
 
 import javax.swing.BorderFactory;
+import javax.swing.ButtonGroup;
 import javax.swing.JButton;
 import javax.swing.JCheckBox;
+import javax.swing.JFileChooser;
 import javax.swing.JLabel;
+import javax.swing.JOptionPane;
 import javax.swing.JPanel;
+import javax.swing.JRadioButton;
 import javax.swing.JTextField;
 
+import net.sf.jabref.Globals;
 import net.sf.jabref.JabRefPreferences;
 import net.sf.jabref.external.ExternalFileTypeEditor;
 import net.sf.jabref.external.push.PushToApplication;
@@ -45,6 +58,13 @@ class ExternalTab extends JPanel implements PrefsTab {
     private final JTextField citeCommand;
     private final JCheckBox openFoldersOfAttachedFiles;
 
+    private final JRadioButton defaultConsole;
+    private final JRadioButton specifiedConsole;
+    private final JTextField consoleEmulatorPath;
+    private final JFileChooser consoleChooser;
+    private final JButton browseButton;
+
+
     public ExternalTab(JabRefFrame frame, PreferencesDialog prefsDiag, JabRefPreferences prefs) {
         this.prefs = prefs;
         this.frame = frame;
@@ -53,6 +73,39 @@ class ExternalTab extends JPanel implements PrefsTab {
         JButton editFileTypes = new JButton(Localization.lang("Manage external file types"));
         citeCommand = new JTextField(25);
         editFileTypes.addActionListener(ExternalFileTypeEditor.getAction(prefsDiag));
+
+        defaultConsole = new JRadioButton(Localization.lang("Use default terminal emulator"));
+        specifiedConsole = new JRadioButton(Localization.lang("Specify terminal emulator") + ":");
+        consoleEmulatorPath = new JTextField(25);
+        consoleChooser = new JFileChooser();
+        browseButton = new JButton(Localization.lang("Browse"));
+
+        ButtonGroup consoleOptions = new ButtonGroup();
+        consoleOptions.add(defaultConsole);
+        consoleOptions.add(specifiedConsole);
+
+        JPanel consoleOptionPanel = new JPanel(new GridBagLayout());
+        GridBagConstraints layoutConstraints = new GridBagConstraints();
+
+        defaultConsole.addActionListener(new ConsoleRadioButtonActionListener());
+        specifiedConsole.addActionListener(new ConsoleRadioButtonActionListener());
+        browseButton.addActionListener(new BrowseButtonActionListener());
+
+        layoutConstraints.fill = GridBagConstraints.HORIZONTAL;
+
+        layoutConstraints.gridx = 0;
+        layoutConstraints.gridy = 0;
+        layoutConstraints.insets = new Insets(0, 0, 6, 0);
+        consoleOptionPanel.add(defaultConsole, layoutConstraints);
+
+        layoutConstraints.gridy = 1;
+        consoleOptionPanel.add(specifiedConsole, layoutConstraints);
+
+        layoutConstraints.gridx = 1;
+        consoleOptionPanel.add(consoleEmulatorPath, layoutConstraints);
+        layoutConstraints.gridx = 2;
+        layoutConstraints.insets = new Insets(0, 4, 6, 0);
+        consoleOptionPanel.add(browseButton, layoutConstraints);
 
         FormLayout layout = new FormLayout(
                 "1dlu, 8dlu, left:pref, 4dlu, fill:150dlu, 4dlu, fill:pref", "");
@@ -92,6 +145,12 @@ class ExternalTab extends JPanel implements PrefsTab {
         builder.nextLine();
         builder.append(pan);
         builder.append(editFileTypes);
+        builder.nextLine();
+
+        builder.appendSeparator(Localization.lang("Open console"));
+        builder.nextLine();
+        builder.append(new JPanel());
+        builder.append(consoleOptionPanel);
 
         pan = builder.getPanel();
         pan.setBorder(BorderFactory.createEmptyBorder(5, 5, 5, 5));
@@ -113,6 +172,12 @@ class ExternalTab extends JPanel implements PrefsTab {
         openFoldersOfAttachedFiles.setSelected(prefs.getBoolean(JabRefPreferences.OPEN_FOLDERS_OF_ATTACHED_FILES));
 
         citeCommand.setText(prefs.get(JabRefPreferences.CITE_COMMAND));
+
+        defaultConsole.setSelected(Globals.prefs.getBoolean(JabRefPreferences.USE_DEFAULT_CONSOLE_APPLICATION));
+        specifiedConsole.setSelected(!Globals.prefs.getBoolean(JabRefPreferences.USE_DEFAULT_CONSOLE_APPLICATION));
+        consoleEmulatorPath.setText(Globals.prefs.get(JabRefPreferences.CONSOLE_APPLICATION));
+
+        updateEnableStatus();
     }
 
     @Override
@@ -120,15 +185,53 @@ class ExternalTab extends JPanel implements PrefsTab {
         prefs.put(JabRefPreferences.EMAIL_SUBJECT, emailSubject.getText());
         prefs.putBoolean(JabRefPreferences.OPEN_FOLDERS_OF_ATTACHED_FILES, openFoldersOfAttachedFiles.isSelected());
         prefs.put(JabRefPreferences.CITE_COMMAND, citeCommand.getText());
+        prefs.putBoolean(JabRefPreferences.USE_DEFAULT_CONSOLE_APPLICATION, defaultConsole.isSelected());
+        prefs.put(JabRefPreferences.CONSOLE_APPLICATION, consoleEmulatorPath.getText());
     }
 
     @Override
     public boolean validateSettings() {
+        if (!consoleEmulatorPath.getText().trim().isEmpty()) {
+            Path path = Paths.get(consoleEmulatorPath.getText());
+
+            if (!Files.exists(path) || Files.isDirectory(path) || !path.isAbsolute()) {
+                JOptionPane.showMessageDialog(null,
+                        Localization.lang("Please type in the absolute path to an existing terminal emulator."),
+                        Localization.lang("Error"), JOptionPane.ERROR_MESSAGE);
+                return false;
+            }
+        }
         return true;
     }
 
     @Override
     public String getTabName() {
         return Localization.lang("External programs");
+    }
+
+
+    private class BrowseButtonActionListener implements ActionListener {
+
+        @Override
+        public void actionPerformed(ActionEvent e) {
+            int answer = consoleChooser.showOpenDialog(ExternalTab.this);
+            if (answer == JFileChooser.APPROVE_OPTION) {
+                consoleEmulatorPath.setText(consoleChooser.getSelectedFile().getAbsolutePath());
+            }
+        }
+    }
+
+    private class ConsoleRadioButtonActionListener implements ActionListener {
+
+        @Override
+        public void actionPerformed(ActionEvent e) {
+            updateEnableStatus();
+        }
+    }
+
+
+    private void updateEnableStatus() {
+        consoleEmulatorPath.setEnabled(specifiedConsole.isSelected());
+        browseButton.setEnabled(specifiedConsole.isSelected());
     }
 }
